Update ndrFullDeviceCollectorReport
This report will return devices that are managed by Sophos and unmanaged. The unmanaged report will show its IP and MAC address with a suggestion for identifying the hardware. The managed report will include IP, MAC address, Hostname, OS Name, and current Version. 

This will report on an environment with multiple collectors and sort by the collector ID and group by MAC address.

NOTE: if using a VLAN enabled device, it's possible that you'll have duplicated IPs on a specific MAC address.